+++ SignNow.Net/Service/EventSubscriptionService.cs
@@ -174,5 +174,26 @@ namespace SignNow.Net.Service
                 .RequestAsync<EventHistoryListResponse>(requestOptions, cancellationToken)
                 .ConfigureAwait(false);
         }
+
+        /// <inheritdoc />
+        public async Task<CallbacksResponse> GetCallbacksAsync(GetCallbacksOptions options = default, CancellationToken cancellationToken = default)
+        {
+            Token.TokenType = TokenType.Bearer;
+
+            var query = options?.ToQueryString();
+            var filters = string.IsNullOrEmpty(query)
+                ? string.Empty
+                : $"?{query}";
+
+            var requestOptions = new GetHttpRequestOptions
+            {
+                RequestUrl = new Uri(ApiBaseUrl, $"/v2/event-subscriptions/callbacks{filters}"),
+                Token = Token
+            };
+
+            return await SignNowClient
+                .RequestAsync<CallbacksResponse>(requestOptions, cancellationToken)
+                .ConfigureAwait(false);
+        }
     }
 }